Page element list, element attributes, etc.
When the page loads or a new element is added, the list of elements on the right-side menu is updated. Below that list is the list of css attributes of any selected element. Also, made a small refactor in index, and made little aesthetic changes to the right-side menu and file browser.