Basic creation of elements
The functions I'm making are hard-coded, but that may change. For now, dragging a function from the list adds it as the element is was created to be, so the "image" function makes an image, and a "div" function makes a div. Also, some work begun on save/ load functionality for text editor.